Update index.html
### Acknowledgement
This material is based upon work supported by the National Science Foundation under Grant No. 
[IIS-1546083](http://nsf.gov/awardsearch/showAward?AWD_ID=1546083) (2016-2019), "Holistic Optimization of Data-Driven Applications,"
and is a collaborative effort between 
the [University of Washington](http://db.cs.washington.edu) and the [University of Chicago](http://systems.cs.uchicago.edu).

Any opinions, findings, and conclusions or recommendations expressed in this
material are those of the author(s) and do not necessarily reflect the views of
the National Science Foundation.